Move automatic SEV generation autodetection into sev
This will make it easier to run the SEV launch test on any machine.

No functional changes.
